simplify number of imports needed when importing keypairs
