+++ .changeset/fresh-drinks-itch.md
@@ -1,0 +1,5 @@
+---
+'@mysten/sui': patch
+---
+
+Reduce number of imports required when importing keypairs
