+++ packages/typescript/src/cryptography/publickey.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import { toBase64 } from '@mysten/bcs';
+import { fromBase64, toBase64 } from '@mysten/bcs';
 import { blake2b } from '@noble/hashes/blake2b';
 import { bytesToHex } from '@noble/hashes/utils';
 
@@ -9,6 +9,7 @@ import { bcs } from '../bcs/index.js';
 import { normalizeSuiAddress, SUI_ADDRESS_LENGTH } from '../utils/sui-types.js';
 import type { IntentScope } from './intent.js';
 import { messageWithIntent } from './intent.js';
+import { SIGNATURE_FLAG_TO_SCHEME, SIGNATURE_SCHEME_TO_SIZE } from './signature-scheme.js';
 
 /**
  * Value to be converted into public key.
@@ -137,4 +138,31 @@ export abstract class PublicKey {
 	 * Verifies that the signature is valid for for the provided message
 	 */
 	abstract verify(data: Uint8Array, signature: Uint8Array | string): Promise<boolean>;
+}
+
+export function parseSerializedKeypairSignature(serializedSignature: string) {
+	const bytes = fromBase64(serializedSignature);
+
+	const signatureScheme =
+		SIGNATURE_FLAG_TO_SCHEME[bytes[0] as keyof typeof SIGNATURE_FLAG_TO_SCHEME];
+
+	switch (signatureScheme) {
+		case 'ED25519':
+		case 'Secp256k1':
+		case 'Secp256r1':
+			const size =
+				SIGNATURE_SCHEME_TO_SIZE[signatureScheme as keyof typeof SIGNATURE_SCHEME_TO_SIZE];
+			const signature = bytes.slice(1, bytes.length - size);
+			const publicKey = bytes.slice(1 + signature.length);
+
+			return {
+				serializedSignature,
+				signatureScheme,
+				signature,
+				publicKey,
+				bytes,
+			};
+		default:
+			throw new Error('Unsupported signature scheme');
+	}
 }
