+++ packages/typescript/src/cryptography/signature.ts
@@ -7,13 +7,10 @@ import { bcs } from '../bcs/index.js';
 import { parseSerializedPasskeySignature } from '../keypairs/passkey/publickey.js';
 import type { MultiSigStruct } from '../multisig/publickey.js';
 import { parseSerializedZkLoginSignature } from '../zklogin/publickey.js';
+import { parseSerializedKeypairSignature } from './publickey.js';
 import type { PublicKey } from './publickey.js';
 import type { SignatureScheme } from './signature-scheme.js';
-import {
-	SIGNATURE_FLAG_TO_SCHEME,
-	SIGNATURE_SCHEME_TO_FLAG,
-	SIGNATURE_SCHEME_TO_SIZE,
-} from './signature-scheme.js';
+import { SIGNATURE_FLAG_TO_SCHEME, SIGNATURE_SCHEME_TO_FLAG } from './signature-scheme.js';
 
 /**
  * Pair of signature and corresponding public key
@@ -65,24 +62,14 @@ export function parseSerializedSignature(serializedSignature: string) {
 				signatureScheme,
 				multisig,
 				bytes,
+				signature: undefined,
 			};
 		case 'ZkLogin':
 			return parseSerializedZkLoginSignature(serializedSignature);
 		case 'ED25519':
 		case 'Secp256k1':
 		case 'Secp256r1':
-			const size =
-				SIGNATURE_SCHEME_TO_SIZE[signatureScheme as keyof typeof SIGNATURE_SCHEME_TO_SIZE];
-			const signature = bytes.slice(1, bytes.length - size);
-			const publicKey = bytes.slice(1 + signature.length);
-
-			return {
-				serializedSignature,
-				signatureScheme,
-				signature,
-				publicKey,
-				bytes,
-			};
+			return parseSerializedKeypairSignature(serializedSignature);
 		default:
 			throw new Error('Unsupported signature scheme');
 	}
