+++ packages/typescript/src/keypairs/ed25519/publickey.ts
@@ -4,10 +4,13 @@
 import { fromBase64 } from '@mysten/bcs';
 import { ed25519 } from '@noble/curves/ed25519';
 
+import {
+	bytesEqual,
+	parseSerializedKeypairSignature,
+	PublicKey,
+} from '../../cryptography/publickey.js';
 import type { PublicKeyInitData } from '../../cryptography/publickey.js';
-import { bytesEqual, PublicKey } from '../../cryptography/publickey.js';
 import { SIGNATURE_SCHEME_TO_FLAG } from '../../cryptography/signature-scheme.js';
-import { parseSerializedSignature } from '../../cryptography/signature.js';
 
 const PUBLIC_KEY_SIZE = 32;
 
@@ -67,7 +70,7 @@ export class Ed25519PublicKey extends PublicKey {
 	async verify(message: Uint8Array, signature: Uint8Array | string): Promise<boolean> {
 		let bytes;
 		if (typeof signature === 'string') {
-			const parsed = parseSerializedSignature(signature);
+			const parsed = parseSerializedKeypairSignature(signature);
 			if (parsed.signatureScheme !== 'ED25519') {
 				throw new Error('Invalid signature scheme');
 			}
