+++ packages/typescript/src/keypairs/secp256k1/publickey.ts
@@ -5,10 +5,13 @@ import { fromBase64 } from '@mysten/bcs';
 import { secp256k1 } from '@noble/curves/secp256k1';
 import { sha256 } from '@noble/hashes/sha256';
 
-import { bytesEqual, PublicKey } from '../../cryptography/publickey.js';
+import {
+	bytesEqual,
+	parseSerializedKeypairSignature,
+	PublicKey,
+} from '../../cryptography/publickey.js';
 import type { PublicKeyInitData } from '../../cryptography/publickey.js';
 import { SIGNATURE_SCHEME_TO_FLAG } from '../../cryptography/signature-scheme.js';
-import { parseSerializedSignature } from '../../cryptography/signature.js';
 
 const SECP256K1_PUBLIC_KEY_SIZE = 33;
 
@@ -68,7 +71,7 @@ export class Secp256k1PublicKey extends PublicKey {
 	async verify(message: Uint8Array, signature: Uint8Array | string): Promise<boolean> {
 		let bytes;
 		if (typeof signature === 'string') {
-			const parsed = parseSerializedSignature(signature);
+			const parsed = parseSerializedKeypairSignature(signature);
 			if (parsed.signatureScheme !== 'Secp256k1') {
 				throw new Error('Invalid signature scheme');
 			}
